requirements.txt
wheel
python-telegram-bot==13.11
responses
pytest==5.4.3
pyyaml
numpy
matplotlib==3.2.2
